fixed bug connected to platform differnce... there was hardcoded uint32_t in frame, which didnt work on my wsl due to size_t being 8 bytes

+++ src/frame/frame.cpp
@@ -3,7 +3,6 @@
 #include "transport.h"
 
 #include <cstddef>
-#include <cstdint>
 #include <memory>
 #include <optional>
 #include <stdexcept>
@@ -19,7 +18,7 @@ void Framer::send_message(const buffer& message) {
 }
 
 std::optional<buffer> Framer::recv_message() {
-    uint32_t size = 0;
+    size_t size = 0;
 
     try {
         read_all(reinterpret_cast<std::byte*>(&size), sizeof(size));
